ARM64: Some FP work.

+++ Core/MIPS/ARM64/Arm64RegCache.cpp
@@ -68,7 +68,7 @@ bool Arm64RegCache::IsMapped(MIPSGPReg mipsReg) {
 	return mr[mipsReg].loc == ML_ARMREG;
 }
 
-void Arm64RegCache::SetRegImm(ARM64Reg reg, u32 imm) {
+void Arm64RegCache::SetRegImm(ARM64Reg reg, u64 imm) {
 	// On ARM64, at least Cortex A57, good old MOVT/MOVW  (MOVK in 64-bit) is really fast.
 	emit_->MOVI2R(reg, imm);
 }
